Use pathlib to make CSV exporting cross-platform
