+++ sroka/api/mysql/mysql.py
@@ -1,12 +1,11 @@
-from configparser import NoOptionError, NoSectionError
-
 import os
 import mysql.connector
-from mysql.connector.errors import DatabaseError, OperationalError, InternalError
 import pandas as pd
+from configparser import NoSectionError
+from pathlib import Path
+from mysql.connector.errors import DatabaseError, OperationalError, InternalError
 from retrying import retry
-
-import sroka.config.config as config
+from sroka.api.mysql.mysql_helpers import validate_options, get_options_from_config
 
 
 @retry(stop_max_attempt_number=1,
@@ -52,73 +51,26 @@ def query_mysql(query: str, filename=None):
     cursor.close()
     connection.close()
 
-    # Get path to folder in which to create the file.
-    filepath = "\\".join(filename.split('\\')[:-1])
-
-    # If it does not exist, create it.
-    if not os.path.exists(filepath):
-        os.makedirs(filepath, exist_ok=True)
-
-    # Save data to CSV file if a filename was provided, return it in a pandas
-    # dataframe otherwise.
-    if filename:
-        df.to_csv(filename)
-    else:
+    # If no filename is specified, return the data as a pandas Dataframe.
+    # Otherwise, store it in a file.
+    if not filename:
         return df
 
+    # Store the path in a cross-platform pathlib object to ensure compatibility
+    # with DOS & UNIX-based operating systems.
+    path = Path(filename)
 
-def get_options_from_config():
-    # Set the options in a dictionary, in order to pass only the
-    # options that were provided in the configuration file to the
-    # MySQL connector. Passing empty values would trigger exceptions.
-    options = dict()
+    # Get the parent directory of the given path, if it exists.
+    directory_path = str(path.parent.resolve())
 
+    # If the given path points to a folder, attempt to create it. If it already
+    # exists, the `exist_ok` option ensures that no exception will be thrown.
+    if directory_path != "":
+        os.makedirs(directory_path, exist_ok=True)
+
+    # Export the data in a CSV file.
     try:
-        options["host"] = config.get_value('mysql', 'host')
-    except (KeyError, NoOptionError):  # Do nothing, this value is optional.
-        pass
+        df.to_csv(filename)
+    except OSError as e:
+        print('Unable to write on filesystem: {}'.format(e))
 
-    try:
-        options["port"] = config.get_value('mysql', 'port')
-    except (KeyError, NoOptionError):  # Do nothing, this value is optional.
-        pass
-
-    try:
-        options["user"] = config.get_value('mysql', 'user')
-    except (KeyError, NoOptionError):  # Do nothing, this value is optional.
-        pass
-
-    try:
-        options["password"] = config.get_value('mysql', 'password')
-    except (KeyError, NoOptionError):  # Do nothing, this value is optional.
-        pass
-
-    try:
-        options["unix_socket"] = config.get_value('mysql', 'unix_socket')
-    except (KeyError, NoOptionError):  # Do nothing, this value is optional.
-        pass
-
-    try:
-        options["database"] = config.get_value('mysql', 'database')
-    except (KeyError, NoOptionError):  # Do nothing, this value is optional.
-        pass
-
-    return options
-
-
-# Ensures that either the options are set to connect to a remote database
-# through a host/port combination, or through a unix socket.
-def validate_options(options: dict):
-    # If no connection option is given, there is not enough information to
-    # connect to the database.
-    if "unix_socket" not in options and "host" not in options:
-        print('Invalid Configuration: In order to connect to the MySQL database, the host/port options or the unix_socket option must be set.')
-        return False
-
-    # If both keys are present and both have values, the configuration is
-    # ambiguous.
-    if "unix_socket" in options and "host" in options and \
-        options["unix_socket"] != "" and options["host"] != "":
-        print('Invalid Configuration: In order to connect to the MySQL database, the host/port options or the unix_socket option must be set.')
-        return False
-    return True
